Rename files -controller to .controller, -routes to .routes and fix its imports
