+++ src/modules/posts/index.ts
@@ -1,4 +1,4 @@
-import postsRoutes from './posts-routes';
+import postsRoutes from './posts.routes';
 
 export default {
   postsRoutes
